requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.3
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.1
Pillow==8.0.1
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
typed-ast==1.4.1
typing-extensions==3.7.4.3
